feat: set primary phone and email while validating contact (LAN-200)
