+++ landa/hooks.py
@@ -278,7 +278,10 @@ doc_events = {
 		"autoname": "landa.organization_management.address.address.autoname",
 	},
 	"Contact": {
-		"validate": "landa.address_and_contact.validate",
+		"validate": [
+			"landa.organization_management.contact.contact.validate",
+			"landa.address_and_contact.validate"
+		],
 		"after_insert": "landa.organization_management.contact.contact.after_insert",
 	},
 }
